Add a check to see if Jenkins is clobbering tags
If the git tag begins with "jenkins", there's a pretty strong chance
that Jenkins is clobbering our git tag. I've added logic that checks for
a new env var, `semverIfJenkins`. If that is set to 1 when ldflags.sh is
called, ldflags.sh will check to see if the git tag begins with
"jenkins", and, if so, it grabs the latest git tag beginning with the
character "v" assuming that is a semver tag.

If semverIfJenkins is set but there is no semver tag, gittag will be
empty string.